Add socket-to-PID mapping infrastructure
Created a HashMap to store socket cookie -> PID mappings. The TC program
now attempts to look up the PID for each packet based on its socket cookie.
Currently the map is empty (all lookups will return 'unknown'), but this
lays the groundwork for the cgroup program that will populate it.